+++ src/main/scala/sc4pac/DesktopOps.scala
@@ -10,14 +10,14 @@ object DesktopOps {
     import java.awt.Desktop
     if (Desktop.isDesktopSupported() && Desktop.getDesktop().isSupported(Desktop.Action.BROWSE)) {
       Desktop.getDesktop().browse(url)
-    }
+    } else throw new UnsupportedOperationException("Desktop operations are not supported on this platform.")
   }
 
-  def openDirectory(directory: java.io.File): Task[Unit] = ZIO.attempt {
+  def openDirectory(directory: os.Path): Task[Unit] = ZIO.attempt {
     import java.awt.Desktop
     if (Desktop.isDesktopSupported() && Desktop.getDesktop().isSupported(Desktop.Action.OPEN)) {
-      Desktop.getDesktop().open(directory)
-    }
+      Desktop.getDesktop().open(directory.toIO)
+    } else throw new UnsupportedOperationException("Desktop operations are not supported on this platform.")
   }
 
 }